fix(#2175 V2-S1): materialized __typeof function arm + shared closure classifier
Add a function-classifier arm to the standalone materialized __typeof
native so a closure/function value read back dynamically (const t =
typeof x) reports "function", matching the inline __typeof_function
predicate (#1896). Previously it fell through to "object" — the #2984
path-dependent typeof defect.

Consolidate the duplicated closure-base-wrapper list (index.ts +
dyn-read.ts) into one leaf module closure-classifier.ts
(collectClosureBaseWrapperTypeIdxs + buildClosureRefTestArms) so
__typeof/__typeof_function/__typeof_object/the .length arm share ONE
predicate — never two divergent arm lists (#2949 slice-3 consumer).

✓
